fix(cdn-check): PR feedback, IP metadata order, Franklin markup, tests
- Document fcors key and justify file-level eslint-disable for no-await-in-loop
- Remove debug console logging; keep logProdNetworkDiag for failures
- Prefer RDAP before ipwho/ipapi for IP metadata (fewer ipwho 403s in Network)
- Use section roots for score, report heading, and error UI to avoid phantom blocks
- Note mobile-first breakpoints in cdn-check.css header
- Extract classifyIpNetwork to cdn-check-network.js; add node:test coverage
